+++ src/super.c
@@ -1554,6 +1554,9 @@ double novas_hms_hours(const char *restrict hms) {
  * (arc)second components may be separated by spaces, tabs, colons `:`, underscore `_`, or a
  * combination thereof. Additionally, the degree and minutes may be semarated by the letter `d`
  * or `D`, and the minutes and seconds may be separated by `m` or `M`, or a single quote `'`.
+ * The last component may also be followed by a standalone upper-case letter 'N', 'E', 'S',
+ * or 'W' signifying a compass direction.
+ *
  * For example, all of the lines below specify the same angle:
  *
  * <pre>
@@ -1561,6 +1564,8 @@ double novas_hms_hours(const char *restrict hms) {
  *  -179 59m 59.999
  *  -179d 59' 59.999
  *  -179D59'59.999
+ *  179:59:59.999W
+ *  179 59 59.999 S
  * </pre>
  *
  *
@@ -1581,8 +1586,9 @@ double novas_hms_hours(const char *restrict hms) {
 double novas_parse_dms(const char *restrict dms, char **restrict tail) {
   static const char *fn = "novas_dms_degrees";
 
-  int d = 0, m = 0, n = 0;
+  int d = 0, m = 0, n = 0, k = 0;
   double s = NAN;
+  char compass[3] = {};
 
   if(tail)
     *tail = (char *) dms;
@@ -1611,26 +1617,48 @@ double novas_parse_dms(const char *restrict dms, char **restrict tail) {
     return NAN;
   }
 
+  s = abs(d) + (m / 60.0) + (s / 3600.0);
+  if (d < 0)
+    s = -s;
+
+  if(dms[n-1] == 'E') {
+    // An 'E' immediately after the last numerical value, is parsed as part of the number
+    // but we should interpret it as a compass direction.
+    n++;
+  }
+  else if(sscanf(&dms[n], "%2s%n", compass, &k) > 0) {
+    if(strcmp(compass, "N") == 0 || strcmp(compass, "E") == 0) {
+      n += k;
+    }
+    else if (strcmp(compass, "S") == 0 || strcmp(compass, "W") == 0) {
+      s = -s;
+      n += k;
+    }
+  }
+
   if(tail)
     *tail += n;
 
-  s = abs(d) + (m / 60.0) + (s / 3600.0);
-
-  return d < 0 ? -s : s;
+  return s;
 }
 
 /**
  * Returns the decimal degrees for a DMS string specification. The degree, (arc)minute, and
  * (arc)second components may be separated by spaces, tabs, colons `:`, or a combination thereof.
  * Additionally, the degree and minutes may be semarated by the letter `d` or `D`, and the
- * minutes and seconds may be separated by `m` or `M`, or a single quote `'`. For example, all of
- * the lines below specify the same angle:
+ * minutes and seconds may be separated by `m` or `M`, or a single quote `'`. The last component
+ * may also be followed by a standalone upper-case letter 'N', 'E', 'S', or 'W' signifying a
+ * compass direction.
+ *
+ * For example, all of the lines below specify the same angle:
  *
  * <pre>
  *  -179:59:59.999
  *  -179 59m 59.999
  *  -179d 59' 59.999
  *  -179D59'59.999
+ *  179:59:59.999S
+ *  179 59 59.999 W
  * </pre>
  *
  *